django_project2/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-debug-toolbar==5.2.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
